+++ src/Compilers/Core/Rebuild/CompilationFactory.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
@@ -12,6 +11,7 @@ using System.Reflection.PortableExecutable;
 using System.Threading;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Emit;
+using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.Rebuild
@@ -136,6 +136,9 @@ namespace Microsoft.CodeAnalysis.Rebuild
                 pdbFilePath = codeView.Path ?? throw new InvalidOperationException("Could not get PDB file path");
             }
 
+            var rebuildData = new RebuildData(
+                OptionsReader.GetMetadataCompilationOptionsBlobReader(),
+                getNonSourceFileDocumentNames(OptionsReader.PdbReader, OptionsReader.GetSourceFileCount()));
             var emitResult = rebuildCompilation.Emit(
                 peStream: rebuildPeStream,
                 pdbStream: rebuildPdbStream,
@@ -150,12 +153,25 @@ namespace Microsoft.CodeAnalysis.Rebuild
                     subsystemVersion: SubsystemVersion.Create(peHeader.MajorSubsystemVersion, peHeader.MinorSubsystemVersion)),
                 debugEntryPoint: debugEntryPoint,
                 metadataPEStream: null,
-                rebuildData: new RebuildData(OptionsReader.GetMetadataCompilationOptionsBlobReader(), OptionsReader.PdbReader, OptionsReader.GetSourceFileCount()),
+                rebuildData: rebuildData,
                 sourceLinkStream: sourceLink != null ? new MemoryStream(sourceLink) : null,
                 embeddedTexts: embeddedTexts,
                 cancellationToken: cancellationToken);
 
             return emitResult;
+
+            static ImmutableArray<string> getNonSourceFileDocumentNames(MetadataReader pdbReader, int sourceFileCount)
+            {
+                var count = pdbReader.Documents.Count - sourceFileCount;
+                var builder = ArrayBuilder<string>.GetInstance(count);
+                foreach (var documentHandle in pdbReader.Documents.Skip(sourceFileCount))
+                {
+                    var document = pdbReader.GetDocument(documentHandle);
+                    var name = pdbReader.GetString(document.Name);
+                    builder.Add(name);
+                }
+                return builder.ToImmutableAndFree();
+            }
 
             IMethodSymbol? getDebugEntryPoint()
             {
